Abstracted Language Selector to seperate component
